range, range6: bound the lease table and get sqlite off the plugin lock
Every fresh DUID and IAID pair used to get a binding until the pool ran out,
which on a /96 means the heap and the lease file go first, so both plugins
now take max-leases (65536 by default, 0 to turn it off), count what they
loaded at startup against it and refuse new leases past it with a paced log
line, reclaiming what has lapsed before they turn anyone away. Writes moved
onto one writer goroutine per instance that applies changes in the order the
packet path made them, so the lock is no longer held across a sqlite write;
the queue is bounded, and a lease that cannot be queued is refused rather
than handed out with no row behind it. Expiry is stored as int64 so a 32-bit
build does not read every lease as expired in 2038, storage runs under the
plugin's own context with a timeout per statement, failures come back as
ErrNotFound, ErrBusy, ErrCorruptRecord or ErrWriteQueueFull, and range
filters and truncates the client hostname the way range6 already did.

+++ plugins/range6/leases.go
@@ -55,7 +55,7 @@ func (p *pluginState) Leases() []leases.Lease {
 			// Expiry is stored as a Unix second. It goes out in UTC because
 			// an API answer is usually read somewhere other than the host
 			// that produced it.
-			Expires:  time.Unix(int64(record.expires), 0).UTC(),
+			Expires:  time.Unix(record.expires, 0).UTC(),
 			Address:  netip.PrefixFrom(addr, addr.BitLen()),
 			Hostname: record.hostname,
 			Source:   p.name,
